Demo/focus_tracking/requirements.txt
# Requires Python 3.10

mediapipe==0.9.3
opencv-python==4.5.5.64
numpy==1.22.4
Pillow==9.2.0
pyautogui==0.9.53
keyboard==0.13.5